feat(scheduler): move scheduler function into scheduler.js
